cros_bundle_firmware: Support USB download properly on Tegra
Now that we can select between downloading a flasher or a normal U-Boot
with the -F flag, support this on Tegra too.

BUG=chrome-os-partner:10075
TEST=manual:
build for seaboard
$ cros_bundle_firmware -w usb
See that it sends U-Boot to the board and it boots
$ cros_bundle_firmware -w usb -F spi
See that it sends a flasher to the board and it successfully flashes U-Boot.
Boot the new image from SPI and see that it works.

Change-Id: I2b570cbc3b6af5466239a4c16145181e3f75d9a2
Reviewed-on: https://gerrit.chromium.org/gerrit/24765
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>