Add apache setup script to set up an apache server to run with the devserver.
This script and accompany configs sets up apache to run on port 8082 and direct
non-file traffic to the devserver i.e. for any paths not in the static dir.
This means all RPCs will go to the running devserver on 8080 yet apache
will handle requests to files.

Once this script is run. We need only modify the autotest config to contact
the devserver on port 8082 rather than 8080 allowing for easy staging/testing
of apache with an autotest server.

BUG=chromium-os:26451
TEST=Set up locally against my own static dir with the devserver running
in production settings. Made calls to download and wait_for_url and also
ensured I could grab files from static/archive.

Change-Id: Ia1ff2f3da7f038e64d7a577fed340b8b7d7797c0
Reviewed-on: https://gerrit.chromium.org/gerrit/24939
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>