Modify 'willis' to report default branch discrepancies
The willis utility is used to show the state of all repositories of
the project. It fails to display an important state though: when a
certain repo is on 'no branch' state, but the top SHA1 does not match
the tracking branch. (This could happen if one, say, was trying to
bisect a problem in a repo, checking out different revisions without
creating a branch).

This change modifies willis behavior such that in the case when a
default repo state does not match the tracking branch, this repo is
included in the willis report with a prefix of '!!!!!' (on red
background where appropriate).

Using the chromite library provides easy access to a host of
information about git repositories as well as to the list of
repositories included in the project.

Some effort has been taken to make sure that willis can be invoked
outside of chroot also: if the original attempt to import chromite
library fails, another attempt is made with a corrected Python include
path. The search for the top of the tree is now originating in the
directory where willis is located, this is handy when inwoking outside
of chroot and outside of the tree.

Yet another modification is handing the case where the directory does
not yet exists in the source tree but is present in the manifest.

BUG=none
TEST=manual
  . modify one of the ChromeOS project repositories by checking out a
    random branch, say run

    git checkout HEAD~2

  . run the original version of willis, observe results (the modified
    repo is not included in the output)
  . run modified version of willis, observe results (the modified repo
    is included in the output prefixed with !!!!! on read background

 . run modified version of willis from outside of chroot, invoking it
   from the user home directory, observe it succeed.

Change-Id: I430c7456e734775aae13e71f94e7471a4a1fafa8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/25237