Fixes for apache set up -- use x-forwarded-host if it exists.
This change modifies the apache configuration and tweaks the update
ping logic to handle being behind apache. When behind apache,
the request.base is re-written to whatever is configured in the [proxy]
stanza.  As this is then 127.0.0.1, we need to checked X-Forwarded-Host
for the original hostname used in the request.

The fixes in setup_apache_setup.sh are minor and caught when staging
apache on the live devserver.

BUG=chromium-os:26451
TEST=Ran apache set up on my own host.  Used DUT and contacted my_ip:8082
and saw the hostname get written correctly and the DUT update.

Change-Id: Id7e2de7d6e3e2c2a5fdbc7e5054cab26fa7d6bbd
Reviewed-on: https://gerrit.chromium.org/gerrit/25262
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Scott Zawalski <scottz@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>