cros_bundle_firmware: Allow SPL to specify what it loads
Rather than hard-code the areas that U-Boot SPL loads (to be U-Boot
plus the FDT) we might as well specify it explicitly. This lets us
easily change it later if needed.

BUG=chromium-os:32034
TEST=manual
$ FEATURES=test emerge-daisy cros-devutils
Run cros_bundle_firmware -b daisy, and see that it still produces an
image that boots, and that the listed SPL payload size looks correct.

Change-Id: Id54f7885cce6529d96c3b2f9b53eb120659bd2fd
Reviewed-on: https://gerrit.chromium.org/gerrit/25864
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>