cros_bundle_firmware: Add option to specify EC binary
The EC (Embedded Controller) is a separate microprocessor in the system which
also has firmware. Allow this as a new file which can be placed in a blob in
the flash map.

BUG=chromium-os:32034
TEST=manual
$ FEATURES=test emerge-daisy cros-devutils
Run cros_bundle_firmware with -e option and see that it completes
successfully.

Change-Id: I44abd16d8a3154f91a20e332f53eb5d0c4ee2c12
Reviewed-on: https://gerrit.chromium.org/gerrit/25861
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>