Update common.sh boilerplate in src/platform/dev to be consistent.
Currently there are several different implementations of the boilerplate in
src/platform/dev, some of which are buggy. I've updated all the
boilerplate to use the same implementation, except for
cros_generate_update_payload, which is a special case.

These scripts now consistently print the following message when
common.sh is not installed (or you are running outside the chroot):
  /usr/lib/crosutils/common.sh: No such file or directory

BUG=chromium-os:32442
TEST=Run these scripts with and without common.sh installed.

Change-Id: Ia4c3d7ea90d543018eefcc0f8f00af071970b157
Reviewed-on: https://gerrit.chromium.org/gerrit/27046
Reviewed-by: Chris Sosa <sosa@chromium.org>
Commit-Ready: David James <davidjames@chromium.org>
Tested-by: David James <davidjames@chromium.org>