cros_bundle_firmware: Provide a default textbase
When the fdt does not specify the textbase, we can read it from the
U-Boot image anyway. So use a default of zero.

BUG=chromium-os:32468
TEST=manual
run cros_bundle_firmware using a device tree file without a textbase
property in chromeos-config. See that it no longer fails.

Change-Id: Ic20f0284226ad8b473e3e5a33c53ed16381262eb
Reviewed-on: https://gerrit.chromium.org/gerrit/27158
Reviewed-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>