cros_bundle_firmware: Detect tegrarcm failure message
With the change from nvflash to tegrarcm the message printed when usb
is not connected is different. Update the code so we will prompt the
user to reset/connect the board.

BUG=chromium-os:32468
TEST=manual
Have a seabord running in U-Boot
$ cros_bundle_firmware
See that it requests that you reset the board into T20 recovery and
waits 10 seconds.

Change-Id: I2a1f4bb2ce4238bf24a32a37abcce121ef3716d7
Reviewed-on: https://gerrit.chromium.org/gerrit/27160
Reviewed-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>