cros_bundle_firmware: Add default flashmap for Tegra
It is convenient to be able to flash an upstream U-Boot, but this does
not have a Chrome OS flashmap. So create a default one in this case,
allowing up to 512KB for BCT/U-Boot.

BUG=chromium-os:32468
TEST=manual
Verified as part of seaboard flashing.
Run cros_bundle_firmware on upstream U-Boot and see that it downloads
to a Seaboard and boot.

Change-Id: If3160ad35dd84bc0673653b84cb05b61c2fe8436
Reviewed-on: https://gerrit.chromium.org/gerrit/27164
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>