cros_bundle_firmware: Show which blobs are being built
Before building the blobs, indicate which ones will be built as a
dev convenience. When one is missing, clean up the error to show
only the names of those that are being built.

BUG=chromium-os:32468
TEST=manual
cros_bundle_firmware -v3
See that it now prints out the list of blobs it is about to build.

Change-Id: I0154a7c0f1acf547d89242d5e6a61125d9190bd4
Reviewed-on: https://gerrit.chromium.org/gerrit/27758
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>