cros_bundle_firmware: Support textbase in new upstream tegra release
Upstream tegra has added an SPL feature, with a 16KB header on the
front of U-Boot. So when searching for the text base, we need to
skip this header. Make this automatic for now.

BUG=chromium-os:32468
TEST=manual
Checkout upstream U-Boot from tegra/next, which includes Allen Martin's
SPL series:
 44fd742 tegra20: Remove armv4t build flags

Use ''cros_bundle_firmware -w usb -u u-boot-dtb-tegra.bin' to write
U-Boot to the board, and see that it detects the text base correctly,
downloads and boots.

Change-Id: I886126affae74d6b91ab9bb70a2f82363bec2e08
Reviewed-on: https://gerrit.chromium.org/gerrit/27979
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>