cros_bundle_firmware: Put the updated fdt in coreboot cbfs
At present we put the original unmodified fdt file into the coreboot
rom. This is fine if cros_bundle_firmware doesn't update it, but it
does. To date these changes have not mattered much, but now we need to
pass the position of the EC binary in, so need the code to do the right
thing.

We can't currently use the ARM approach of just tacking the file onto
the end of the U-Boot section, since we have no guarantee that
coreboot's ELF loader will even load data we put there. We could do
that if we switched coreboot to loading a raw binary U-Boot instead of
an ELF file. But for now, add the fdt file after it has been updated
fully.

BUG=chrome-os-partner:11148
TEST=manual
$ cros_bundle_firmware -u u-boot.bin -d board/chromebook-x86/dts/link.dts
       -I arch/x86/dts -I cros/dts -b link -w em100

Run on link, type 'vboot_test fmap'
See that ecbin fields are non-zero.

Change-Id: Ie183a5fb04211e959d45376b86dc264f453c3c66
Reviewed-on: https://gerrit.chromium.org/gerrit/28174
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>