cros_bundle_firmware: Support a read-only EC binary property
We currently support the read-write EC binary. Add support for the read-only
binary as well so we can put this into the firmware image and update it.

Rename the existing 'ecbin' property to 'ecrw' since it is actually only the
read-write portion.

BUG=chrome-os-partner:12806
TEST=manual
Manually modify the device tree source file in U-Boot to request ecro, and
rebuild U-Boot

		ro-boot {
			label = "u-boot";
			size = <0x9a000>;
			read-only;
			type = "blob boot,dtb,ecrw,ecro";
			required;
		};

$ cros_bundle_firmware -b daisy -d exynos5250-snow -O out

See that the ecro region now appears in the image:

$ fdtdump out/updated.dtb
...
            ecrw {
                reg = <0x0006e894 0x0000a370>;
            };
            boot {
                reg = <0x00000000 0x000698d0>;
            };
            dtb {
                reg = <0x000698d0 0x00004fc3>;
            };
            ecro {
                reg = <0x00078c04 0x0000a500>;
            };

Change-Id: Ia6f0a2a3e9e280546a441727a877e836542aead7
Reviewed-on: https://gerrit.chromium.org/gerrit/30701
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>