devserver: fix wait_for_status hang due to FileLock restriction
Turns out that lockfile.FileLock.release() only succeeds if the lock in
question was locked by the same thread trying to release it. This of
course is not the case with async download by different threads.
Instead, we're using break_lock(), which will remove the lock regardless
of who put it in place.

BUG=chromium-os:34684

TEST=Unit tests; installation and interactive testing of
download/wait_for_status with a live instance.

Change-Id: I6322f7d6116089b164fbb748f7544dd5f74b4496
Reviewed-on: https://gerrit.chromium.org/gerrit/33950
Commit-Ready: Gilad Arnold <garnold@chromium.org>
Reviewed-by: Gilad Arnold <garnold@chromium.org>
Tested-by: Gilad Arnold <garnold@chromium.org>