cros_bundle_firmware: Show the correct ebuild name
When we look up the path of a tool (e.g. cbfstool) and it is not present
we currently display the full path as the name of the ebuild. We should
display 'cbfstool' instead of '/usr/bin/cbfstool'.

BUG=none
TEST=manual
Remove tegrarcm tool and run 'cros_bundle_firmware'
See that it requests that we emerge 'tegrarcm'.

Change-Id: Ibeb55400b2ae58561474e33e9e3805f433cc5903
Reviewed-on: https://gerrit.chromium.org/gerrit/34384
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>