Creating new fw bins with new version strings.
Creating bvi scripts that will search and replace the
given version strings, and save to the expected filenames.

Added a check to ensure that the new bin file sizes remain
the same as the orignal bin file, since we should be replacing
not appending the .testX string to existing versions.

Removed the cmd line option to dir of new bin files, since we
are creating them ourselves.

BUG=none
TEST=manual runs

Change-Id: I1e517f7fa3c4a6a18759e1a7d57368dedd774e39
Reviewed-on: https://gerrit.chromium.org/gerrit/35852
Tested-by: Yusuf Mohsinally <mohsinally@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>