cros_bundle_firmware: Add a 'firmware-type' in each firmware's fdt
We use the same fdt for each firmware type. But it is convenient to pass in
the firmware type so that U-Boot knows which type it is (RO, RW-A, RW-B).
Add slightly modified fdts for each. Flashmaps can use these blobs (dtb-rwa,
dtb-rwb) instead of 'dtb' in the relevant RW sections.

BUG=chrome-os-partner:14545
BRANCH=link
TEST=boot into U-Boot with pre-emable flags set to 0.

See that it indicates RO firmware:

> vboot_test fmap
...
Firmware type: 0

Then change to RW-A firmware:

> vboot_twostop
...

Once in RW-A firmware:

> vboot_test fmap
...
Firmware type: 1

Note: RW-B has not been tested. This can be done by modifying vboot to choose
B instead of A.

Change-Id: I51099e7d7ecbb4b96fd0f252cc5ed09b39ce4770
Reviewed-on: https://gerrit.chromium.org/gerrit/36203
Reviewed-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>