cros_bundle_firmware: Allow setting the write method from cmdline
Add a -M option to set the method used for writing to the board. This
can be 'tegra' or 'exynos' at present. Normally it is in the fdt, but
for upstream U-Boot it is not.

BUG=chromium-os:36112
TEST=manual
Use '-M exynos' and see that it writes an upstream U-Boot to the exynos
board. It does not actually boot on snow due to memory init problems.

Change-Id: I55a9cf47a181293d78453eb6184547343b0f8295
Reviewed-on: https://gerrit.chromium.org/gerrit/37458
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>