cros_bundle_firmware: Add default flashmap for exynos
Provide a set of default flashmaps for boards that we understand. Rather
than hard-code the tegra support in PackFirmware, pass it in from
bundle_firmware, which knows about the board type. This allows us to
support exynos as well as tegra.

The exynos flashmap has three sections: BL1, SPL, U-Boot.

BUG=chromium-os:36112
TEST=manual
Build and boot upstream U-Boot on snow (it doesn't actually work yet, but
the downloads happen, and when upstream U-Boot is far enough advanced,
it will work).

Change-Id: Idd918de75d38f3ffd93c8d7f0008c6c296980c9e
Reviewed-on: https://gerrit.chromium.org/gerrit/37460
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>