If at start, bad state is detected, clean update state don't reboot.
This is really useful when you are running tests on an image with
modifications to how update_engine is running or changes to upstart
scripts. In this case, rebooting the machine would wipe any local
state or benchmark you are running which you absolutely don't want.

BUG=chromium-os:36956
TEST=Ran it ... a lot.

Change-Id: I254b8b9f82fb32c2bc0802389ba1bcfc0ea70e53
Reviewed-on: https://gerrit.chromium.org/gerrit/39224
Tested-by: Chris Sosa <sosa@chromium.org>
Reviewed-by: Jay Srinivasan <jaysri@chromium.org>
Commit-Ready: Chris Sosa <sosa@chromium.org>