Update the blob positions in the RW device trees as well.
The positions of the sub-blobs in a particular region of the FMAP are
recorded in the device tree after they've been appended together and their
relative positions are established. This step was being performed for the RO
device tree, but not the RWs. When booting into RW firmware, the EC position
and size weren't found and the system would reboot into recovery mode.

BUG=chrome-os-partner:8339
TEST=Built and booted a U-Boot based firmware on Link. Before this change,
when RO normal mode was disabled, the RW device tree wouldn't have info on the
positon of the EC RW blob, VbExEcGetExpectedRW would return an error, and the
system would reboot into recovery mode. After this change, the info was there
and the system would boot successfully into ChromeOS.
BRANCH=None

Change-Id: I31040832abc4688cb64e29f71a5346197879c318
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/39242
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>