When building blobs with subcomponents, prefix them with an index structure.
The index structure has a 32 bit count, and then pairs of 32 bit integers
which have the offset and length of each subcomponent. The device tree FMAP
is adjusted to reflect the new offsets of the subcomponents.

BUG=chrome-os-partner:16412
TEST=Built U-Boot with and without this change on Link and verified that it
still booted. As part of debugging, printed out the contents of the index and
the start of the RW firmware from within U-Boot and verified that they were
what we expected.
BRANCH=None

Change-Id: Ib2a4a5da1958b92f866a0b2e5b295415b8661e0f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/39419
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>