cros_bundle_firmware: Support lzo compression of blobs
Add a new property to the blob format specifying the compression type
to use for that blob. Implement lzo compression using lzop.

The entire blob is compressed in one step. The offsets within the blob
refer to the uncompressed data, since the compressed data obviously
cannot be extracted in sections.

BUG=chrome-os-partner:13961
BRANCH=none
TEST=manual
Using new link fdt with compression enabled for the RW boot and vblock
areas, boot U-Boot to a kernel.

Change-Id: I098827103474896c6c64d793c9854b196b73b449
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39877
Reviewed-by: Randall Spangler <rspangler@chromium.org>