cros_bundle_firmware: Support 30KB SPL for Exynos
U-Boot will soon support this, so add support for the new size, rather
than hard-coding the SPL size. We do this by detecting the start of
U-Boot.

BUG=chrome-os-partner:12746
TEST=manual
Build and boot snow to kernel with both the smaller and larger SPL.
See that the SPL size is detected correct, U-Boot boots normally and
the kernel runs to a login prompt.

Change-Id: Ia0da542403e7e62db360fb9c802cf2c98f5fcc78
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/40899
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>