When requested, prefix multipart blobs with an index structure.
An index is requested by having a with_index property in the blob or keyblock
type entries in the fdt fmap. The index structure has a 32 bit count, and then
pairs of 32 bit integers which have the offset and length of each
subcomponent. The device tree FMAP is adjusted to reflect the new offsets of
the subcomponents.

BUG=chrome-os-partner:16412
TEST=Built U-Boot with and without this change on Link and verified that it
still booted. As part of debugging, printed out the contents of the index and
the start of the RW firmware from within U-Boot and verified that they were
what we expected. Built snow firmware with and without this change and
verified that it was byte identical except for a version string which was
repeated a few times. Booted firmware built with this change on snow.
BRANCH=None

Change-Id: Ibe5ea068d63c3217e6d898c1bc0cf484cc76da56
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/41188
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>