Make unittests work when firmware tarball is included in the artifact list.
An earlier change (https://gerrit.chromium.org/gerrit/#/c/40258)
altered the return value of common_util.py:ParsePayloadList() without
making the appropriate modifications to unit tests.

This change fixes the failing unit tests and adds another one to test
the case when the firmware tarball is not present in the set. It also
removes a few of unused variables highlighted by presubmit checks.

BUG=chromium-os:38144
TEST=manual
    . in ~/trunk/src/platform/dev execute

     for t in *unittest.py; do
         echo "******************* $t **************************"
         if ! ./$t; then echo "$t failed \!\!\!"
 	     break
	 fi
     done

    observe all tests succeed.

Change-Id: Ica5b71a7c7e990aeafd01fa861126084dd64904f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/41871
Reviewed-by: David James <davidjames@chromium.org>