cros_bundle_firmware: Search a LZO magic when u-boot is LZO compressed
We look for a 'B reset' instruction which is at the start of u-boot to
locate the offset of u-boot. When u-boot is LZO compressed, this method
doesn't work. We should look for a LZO magic instead.

This CL tries to search both patterns and picks the minimal offsets.

BUG=chromium-os:37724
TEST=manual
Use a FDT in which RO u-boot flashmap has the property: compress = "lzo".
Run cros_bundle_firmware -w -p and check the resulted u-boot-from-image.bin
contains a correct compressed u-boot image.

Change-Id: Id6a4b3a50825b35850b70fae2b42f8889e6b1590
Reviewed-on: https://gerrit.chromium.org/gerrit/42693
Commit-Queue: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Tom Wai-Hong Tam <waihong@chromium.org>