Add a new type of blob for the EC hash.
A new vboot API allows us to precompute the EC hash and pass it in instead of
having to load, verify and hash the EC on each boot. This new blob type lets
us put the hash into the image as data. The hash is calculated with hashlib
like the "hash" property which is added to the device tree.

BUG=chrome-os-partner:17606
TEST=Built depthcharge with this new blob type and verified that the EC hash
was retrieved correctly.
BRANCH=None

Change-Id: I3a17a054e2f12fe720d3e41ad3d0d7e33ec7024a
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/42724
Commit-Queue: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>