cros_bundle_firmware: Improve error message for bad flashmap
Add a bit more detail to this error message to help the user find the
right node.

BUG=none
TEST=FEATURES=test sudo  -E emerge cros-devutils

With an incorrect node in the flashmap such as:

		/* ---- Section: Rewritable VPD 32 KB ---- */
		rw-vpd {
			label = "rw-vpd";
			/* Alignment: 4k (for updating) */
			rexg = <0x003f0000 0x00008000>;
			type = "wiped";
			wipe-value = [ff];
		};

it would previously say:

	Must specify either 'reg' or 'size' in flashnode

but that doesn't really help find where the problem is. With this change
we get:

	Must specify either 'reg' or 'size' in flashnode 'rw-vpd'

Change-Id: I2468414bcad19e179dc7fb106560eb6cabdd7808
Reviewed-on: https://gerrit.chromium.org/gerrit/42939
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>